requirements.txt
crawl4ai>=0.7.8
python-dotenv>=1.0.0
pydantic>=2.0.0
mypy>=1.8.0
